Replaced long long to __int64 for Windows
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@11916 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ xmlBlaster/src/c/util/basicDefs.h
@@ -46,7 +46,8 @@ Note:      The gcc and icc (>=8) both define __GNUC__
   typedef int int32_t;
   typedef short int16_t;
 #elif defined(_WINDOWS)
-  typedef long long int64_t;
+  typedef __int64 int64_t;
+  /*typedef long long int64_t;*/
   typedef int int32_t;
   typedef short int16_t;
 #else
